Fetching User info from Firestore 4 A new hope

+++ app/src/main/java/com/example/healthcared/UI/Singup2.kt
@@ -93,7 +93,7 @@ class Singup2 : AppCompatActivity(), AdapterView.OnItemSelectedListener {
             val newUser = Usuario(name, email, username, password, vegan, veget)
             newUser.height = height.toInt()
             newUser.weight = weight.toInt()
-            newUser.targetSteps = date
+            newUser.targetSteps = date.toInt()
             newUser.gender = spinner
 
             Log.v("SIGNUP MAIL", email)
